integrate task generation
