+++ pages/index.js
@@ -3,7 +3,7 @@ import { useState } from "react";
 import styles from "./index.module.css";
 import Dictaphone from './voice/voice';
 
-import  { mapToClickup } from '../utils/clickup';
+import  { mapToClickup } from '../utils/map-to-clickup';
 
 export default function Home() {
   const [result, setResult] = useState();
@@ -30,6 +30,8 @@ export default function Home() {
 
       setResult(data.result);
       await mapToClickup(data.result.stories);
+
+      alert("Success! Check your ClickUp account.");
 
     } catch(error) {
       console.error(error);
@@ -58,7 +60,7 @@ export default function Home() {
         </form>
         <div className={styles.result}>
           {result?.stories?.map((story, index) => (
-            <div className={styles.story}>
+            <div key={index} className={styles.story}>
               <h2>Story Name: { story?.name }</h2>
               <p>Description: { story?.description }</p>
               { story?.tasks?.map((task, index) => (
